Logcollector: added multiline_pattern log type
This multiline file type aggregates lines using a regex to locate the
first or last line of a message.

There are two match models: with "after" all lines not matching the
regex will be appended to the previous line matching it, with "before"
all lines not matching the regex will be prepended to the next line
matching it.

This commit adds the following log type to localfile:

 - multiline_pattern

This commit adds the following options to localfile:

 - multiline_regex: regex pattern used to match anchor lines;
 - multiline_match: can be "after" or "before" (default): select the
   aggregation mode;
 - multiline_negate: negates the regex pattern.